Button: move ButtonTest to a __tests__ directory
